Refund Invoice Pdf

+++ internal/controller/merchant/merchant_subscription_user_subscription_detail.go
@@ -12,7 +12,7 @@ import (
 )
 
 func (c *ControllerSubscription) UserSubscriptionDetail(ctx context.Context, req *subscription.UserSubscriptionDetailReq) (res *subscription.UserSubscriptionDetailRes, err error) {
-	user := query.GetUserAccountById(ctx, uint64(req.UserId))
+	user := query.GetUserAccountById(ctx, req.UserId)
 	utility.Assert(user != nil, "user not found")
 	one := query.GetLatestActiveOrIncompleteOrCreateSubscriptionByUserId(ctx, req.UserId, _interface.GetMerchantId(ctx))
 	if one != nil {
